Add new dependencies to the Cargo.lock file
A large number of dependencies have been added to the Cargo.lock file in line with the program needs. These dependencies cover multiple packages for a variety of functionalities such as gimli, adler, autocfg, backtrace, etc. This thorough update of the Cargo.lock file ensures that the necessary tools are available to create a robust and efficient Rust application.