Add clarification about `userdata.json` location in README
This commit includes a note in the README file to explain that when running the program from the command line, the `userdata.json` file should be in the same directory from where the command is run. This is important as it might be different from the directory of the executable file.