Refactor JSON file handling in get_json_data function
The code for reading and writing to the JSON file within the get_json_data function has been revised. A more graceful error handling has been added in case there's an error in reading from the file, it will initialize the file with empty JSON object. It will also create the JSON file if it does not exist.