Update data saving and field clearing process
The update improves the process of saving data in the JSON file by allowing both the username and API key to be updated concurrently. In addition, it implements an automatic clearing of the username and API key fields once the data have been saved. Furthermore, the update ensures a reloading of the data after any changes have been made.