Change variables types and rename MyApp struct to DataCarrier
The changes in the function variables from String to &str removed unnecessary data cloning and increased efficiency. The MyApp structure has been renamed to DataCarrier to provide a clearer understanding of its use in the code. The use of different functions to process data has been updated to reflect these changes.