Update initialization of DataCarrier struct
Removed the initial call to Simbrief for the flight plan in the main function. The DataCarrier struct now loads the data through the simulated click of the reload flight plan button. Hence, `departure` and `arrival` are now initialized with empty Strings instead of update_fp() function values. Furthermore, `username_changed` is set to `true` upon creation. This change is to ensure a quick display of the GUI