Refactor log file creation and logging function
The 'std::fs' was added to the main scope to carry out file operations. Instead of creating an empty log file at the start of the program, we now remove any pre-existing log file to ensure a clean slate. In addition, the log function in 'logic.rs' underwent some refactoring for cleaner and efficient code, using the write! macro to write to the log file.